+++ tests/commands/test_template.py
@@ -5,8 +5,8 @@ import mock
 from awscfncli import cfn
 from click.testing import CliRunner
 
-from .boto_mock import MOCK_CONFIG, MOCK_CONFIG_WITH_BODY, mock_config, \
-    mock_config_with_body, mock_cfn_client
+from .boto_mock import mock_cfn_client, mock_config_with_templateurl, \
+    mock_config_with_templatebody
 
 __author__ = 'ray'
 __date__ = '1/14/17'
@@ -24,21 +24,24 @@ Error: Missing argument "config_file".
     assert str(result.output) == message
 
 
-def test_cfn_template_validate(mock_config, mock_config_with_body,
-                               mock_cfn_client):
+def test_cfn_template_validate(
+        mock_config_with_templateurl,
+        mock_config_with_templatebody,
+        mock_cfn_client):
     with mock.patch('boto3.client') as mock_client:
         mock_client.return_value = mock_cfn_client
 
         runner = CliRunner()
-        runner.invoke(cfn, ['template', 'validate', mock_config])
+        runner.invoke(cfn, [
+            'template', 'validate', mock_config_with_templateurl])
 
         mock_cfn_client.validate_template.assert_called_with(
-            TemplateURL=MOCK_CONFIG['TemplateURL']
+            TemplateURL='https://s3.amazonaws.com/example.template'
         )
 
         runner = CliRunner()
-        runner.invoke(cfn, ['template', 'validate', mock_config_with_body])
+        runner.invoke(cfn, ['template', 'validate', mock_config_with_templatebody])
 
         mock_cfn_client.validate_template.assert_called_with(
-            TemplateBody=MOCK_CONFIG_WITH_BODY['TemplateBody']
+            TemplateBody='{"Resource": {}}'
         )